+++ types/utils.d.ts
@@ -1,3 +1,4 @@
 export function toNumber(any: any): number | undefined;
-export function toBoolean(any: any): boolean;
+export function isInteger(any: any): boolean;
+export function toBoolean(any: any): boolean | undefined;
 export const log: any;